Change ARM pages
