+++ src/Menu/menu.json
@@ -307,9 +307,7 @@
     "paths": [
       "/asynchronous-requests-and-task-schedules",
       "/asynchronous-requests-and-task-schedules/schedule-a-task",
-      "/asynchronous-requests-and-task-schedules/v1/schedule-a-task",
       "/asynchronous-requests-and-task-schedules/view-edit-scheduled-tasks",
-      "/asynchronous-requests-and-task-schedules/v1/view-edit-scheduled-tasks",
       "/asynchronous-requests-and-task-schedules/run-an-ad-hoc-request",
       "/asynchronous-requests-and-task-schedules/view-requests"
     ],
@@ -319,16 +317,8 @@
         "path": "/asynchronous-requests-and-task-schedules/schedule-a-task"
       },
       {
-        "name": "Schedule a Task (v1)",
-        "path": "/asynchronous-requests-and-task-schedules/v1/schedule-a-task"
-      },
-      {
         "name": "View/Edit Scheduled Tasks",
         "path": "/asynchronous-requests-and-task-schedules/view-edit-scheduled-tasks"
-      },
-      {
-        "name": "View/Edit Scheduled Tasks (v1)",
-        "path": "/asynchronous-requests-and-task-schedules/v1/view-edit-scheduled-tasks"
       },
       {
         "name": "Run a Request",
