+++ src/components/AsynchronousRequestsAndTaskSchedules/TaskRequest/TaskRequest.tsx
@@ -1,7 +1,6 @@
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
 import { z } from "zod";
-
 import { Button } from "@/components/ui/button";
 import {
   Table,
@@ -33,18 +32,24 @@ import {
 import { useARMContext } from "@/Context/ARMContext/ARMContext";
 import {
   IARMAsynchronousTasksTypes,
-  IAsynchronousRequestsAndTaskSchedulesTypes,
+  IAsynchronousRequestsAndTaskSchedulesTypesV1,
+  IScheduleOnce,
+  ISchedulePropsNonPeriodic,
+  ISchedulePropsPeriodic,
 } from "@/types/interfaces/ARM.interface";
 import { X } from "lucide-react";
+import Schedule from "./Schedule";
+import { useGlobalContext } from "@/Context/GlobalContext/GlobalContext";
+import CustomModal4 from "@/components/CustomModal/CustomModal4";
 
-interface ITaskRequestTypes {
+interface ITaskRequestProps {
   action: string;
   handleCloseModal: () => void;
   user_schedule_name: string;
-  selected?: IAsynchronousRequestsAndTaskSchedulesTypes;
+  selected?: IAsynchronousRequestsAndTaskSchedulesTypesV1;
 }
 
-const TaskRequest: FC<ITaskRequestTypes> = ({
+const TaskRequest: FC<ITaskRequestProps> = ({
   action,
   handleCloseModal,
   user_schedule_name,
@@ -53,6 +58,8 @@ const TaskRequest: FC<ITaskRequestTypes> = ({
   const api = useAxiosPrivate();
   const { getAsyncTasks, getTaskParametersByTaskName, setIsSubmit } =
     useARMContext();
+  const { isOpenScheduleModalV1, setIsOpenScheduleModalV1 } =
+    useGlobalContext();
   const [isLoading, setIsLoading] = useState<boolean>(false);
   const [asyncTaskNames, setAsyncTaskNames] = useState<
     IARMAsynchronousTasksTypes[] | undefined
@@ -60,6 +67,19 @@ const TaskRequest: FC<ITaskRequestTypes> = ({
   const [parameters, setParameters] = useState<Record<string, string | number>>(
     selected?.kwargs || {}
   );
+  const [periodic, setPeriodic] = useState<
+    ISchedulePropsPeriodic | undefined
+  >();
+  const [scheduleType, setScheduleType] = useState<string>(
+    selected?.schedule_type ?? ""
+  );
+  const [schedule, setSchedule] = useState<
+    | ISchedulePropsPeriodic
+    | ISchedulePropsNonPeriodic
+    | IScheduleOnce
+    | undefined
+  >(selected?.schedule);
+
   useEffect(() => {
     const fetchAsyncTasks = async () => {
       try {
@@ -85,114 +105,6 @@ const TaskRequest: FC<ITaskRequestTypes> = ({
     });
   }, [parameters]);
 
-  const FormSchema = z.object({
-    user_schedule_name: z.string(),
-    task_name: z.string(),
-    parameters: z.record(z.union([z.string(), z.number()])),
-    schedule: z.string().or(z.number()),
-  });
-
-  const form = useForm<z.infer<typeof FormSchema>>({
-    resolver: zodResolver(FormSchema),
-    defaultValues: {
-      user_schedule_name: "",
-      task_name: "",
-      parameters: action === "Edit Task Schedule" ? selected?.kwargs : {},
-      schedule: action === "Edit Task Schedule" ? selected?.schedule : "",
-    },
-  });
-
-  const onSubmit = async (data: z.infer<typeof FormSchema>) => {
-    if (data.user_schedule_name !== "ad_hoc" && data.schedule === "0") {
-      toast({
-        title: "Info",
-        description: `Schedule value should be greater than 0.`,
-      });
-      return;
-    }
-
-    const adHocPostData = {
-      task_name: data.task_name,
-      parameters: data.parameters,
-    };
-    const scheduleTaskPostData = {
-      user_schedule_name: data.user_schedule_name,
-      task_name: data.task_name,
-      parameters: data.parameters,
-      schedule: Number(data.schedule),
-    };
-    const updateScheduleTaskPostData = {
-      schedule_minutes: Number(data.schedule),
-      parameters: data.parameters,
-    };
-    try {
-      setIsLoading(true);
-
-      if (action === "Ad Hoc") {
-        const response = await api.post(
-          `/asynchronous-requests-and-task-schedules/create-task-schedule`,
-          adHocPostData
-        );
-        if (response) {
-          toast({
-            title: "Success",
-            description: "Ad-Hoc task created successfully.",
-          });
-        } else {
-          toast({
-            title: "Error",
-            description: "Failed to create task schedule.",
-            variant: "destructive",
-          });
-        }
-      } else if (action === "Schedule A Task") {
-        const response = await api.post(
-          `/asynchronous-requests-and-task-schedules/create-task-schedule`,
-          scheduleTaskPostData
-        );
-        if (response) {
-          toast({
-            title: "Success",
-            description: "Schedule A Task successfully.",
-          });
-        } else {
-          toast({
-            title: "Error",
-            description: "Failed to create task schedule.",
-            variant: "destructive",
-          });
-        }
-      } else if (action === "Edit Task Schedule") {
-        const response = await api.put(
-          `/asynchronous-requests-and-task-schedules/update-task-schedule/${selected?.task_name}/${selected?.redbeat_schedule_name}`,
-          updateScheduleTaskPostData
-        );
-        if (response) {
-          toast({
-            title: "Success",
-            description: "Edit Task successfully.",
-          });
-        } else {
-          toast({
-            title: "Error",
-            description: "Failed to create task schedule.",
-            variant: "destructive",
-          });
-        }
-      }
-    } catch (error) {
-      toast({
-        title: "Error",
-        description: "Failed to create task schedule.",
-        variant: "destructive",
-      });
-    } finally {
-      setIsLoading(false);
-      form.reset();
-      setIsSubmit(Math.random() + 23 * 3000);
-    }
-  };
-
   const handleGetParameters = async (task_name: string) => {
     try {
       setIsLoading(true);
@@ -215,10 +127,94 @@ const TaskRequest: FC<ITaskRequestTypes> = ({
     }
   };
 
+  const FormSchema = z.object({
+    user_schedule_name: z.string(),
+    task_name: z.string(),
+    parameters: z.record(z.union([z.string(), z.number()])),
+  });
+
+  const form = useForm<z.infer<typeof FormSchema>>({
+    resolver: zodResolver(FormSchema),
+    defaultValues: {
+      user_schedule_name: "",
+      task_name: "",
+      parameters: selected?.kwargs ?? {},
+    },
+  });
+
+  const onSubmit = async (data: z.infer<typeof FormSchema>) => {
+    if (!(await form.trigger())) return;
+
+    if (action !== "Edit Task Schedule") {
+      if (
+        (scheduleType !== "IMMEDIATE" && data.user_schedule_name === "") ||
+        data.task_name === "" ||
+        scheduleType === "" ||
+        (scheduleType !== "IMMEDIATE" && !schedule)
+      )
+        return toast({
+          title: "Error",
+          description: "Please fill in all fields",
+          variant: "destructive",
+        });
+    }
+    const payload =
+      action === "Schedule A Task" && scheduleType !== "IMMEDIATE"
+        ? {
+            user_schedule_name: data.user_schedule_name,
+            task_name: data.task_name,
+            parameters: data.parameters,
+            schedule,
+            schedule_type: scheduleType,
+          }
+        : scheduleType === "IMMEDIATE"
+        ? {
+            task_name: data.task_name,
+            parameters: data.parameters,
+            schedule_type: scheduleType,
+          }
+        : {
+            schedule,
+            schedule_type: scheduleType,
+            parameters: data.parameters,
+            redbeat_schedule_name: selected?.redbeat_schedule_name,
+          };
+    // console.log(payload, "payload");
+    try {
+      setIsLoading(true);
+      const res = await (action === "Schedule A Task"
+        ? api.post(
+            "/asynchronous-requests-and-task-schedules/create-task-schedule",
+            payload
+          )
+        : api.put(
+            `/asynchronous-requests-and-task-schedules/update-task-schedule/${selected?.task_name}`,
+            payload
+          ));
+      console.log(res, "res");
+      if (res) toast({ title: "Success", description: `${res.data.message}` });
+    } catch (error) {
+      console.log(error, "er");
+      toast({
+        title: "Error",
+        description: "Failed to process request.",
+        variant: "destructive",
+      });
+    } finally {
+      form.reset();
+      setSchedule(undefined);
+      setScheduleType("");
+      setIsLoading(false);
+      setIsSubmit(Math.random() + 23 * 3000);
+    }
+  };
+
   return (
     <div
       className={`${
-        action === "Edit Task Schedule" ? "" : "w-[50%] mx-auto my-10"
+        action === "Edit Task Schedule"
+          ? ""
+          : "w-[900px] h-[450px] mx-auto my-10 border rounded"
       } `}
     >
       {action === "Edit Task Schedule" && (
@@ -227,10 +223,27 @@ const TaskRequest: FC<ITaskRequestTypes> = ({
           <X onClick={() => handleCloseModal()} className="cursor-pointer" />
         </div>
       )}
+      {isOpenScheduleModalV1 === "Schedule" && (
+        <CustomModal4 className="w-[770px] h-[450px]">
+          <Schedule
+            schedule={selected?.schedule ?? schedule}
+            setSchedule={setSchedule}
+            scheduleType={scheduleType}
+            setScheduleType={setScheduleType}
+            action="Schedule"
+            setIsOpenScheduleModalV1={setIsOpenScheduleModalV1}
+            selected={selected}
+            periodic={periodic}
+            setPeriodic={setPeriodic}
+          />
+        </CustomModal4>
+      )}
+
       <Form {...form}>
         <form onSubmit={form.handleSubmit(onSubmit)} className="space-y-2 p-4">
           <div className="grid grid-cols-2 gap-4">
             {user_schedule_name !== "ad_hoc" &&
+              scheduleType !== "IMMEDIATE" &&
               action === "Schedule A Task" && (
                 <FormField
                   control={form.control}
@@ -287,23 +300,14 @@ const TaskRequest: FC<ITaskRequestTypes> = ({
               />
             )}
             {user_schedule_name !== "Ad Hoc" && (
-              <FormField
-                control={form.control}
-                name="schedule"
-                render={({ field }) => (
-                  <FormItem>
-                    <FormLabel>Schedule</FormLabel>
-                    <FormControl>
-                      <Input
-                        {...field}
-                        required
-                        type="text"
-                        placeholder="Schedule"
-                      />
-                    </FormControl>
-                  </FormItem>
-                )}
-              />
+              <div className="flex flex-col gap-[18px] pt-8">
+                <h3
+                  className="bg-gray-400 rounded p-[7px] border text-white font-bold hover:bg-gray-500 text-center cursor-pointer"
+                  onClick={() => setIsOpenScheduleModalV1("Schedule")}
+                >
+                  Schedule
+                </h3>
+              </div>
             )}
           </div>
           <Table>
