--- src/pages/AsynchronousRequestsAndTaskSchedules/ViewEditScheduledTasksV1.tsx
@@ -1,10 +1,0 @@
-import { ViewEditScheduledTasksTableV1 } from "@/components/AsynchronousRequestsAndTaskSchedules/ViewEditScheduledTasksV1/ViewEditScheduledTasksTableV1";
-
-const ViewEditScheduledTasks = () => {
-  return (
-    <div>
-      <ViewEditScheduledTasksTableV1 />
-    </div>
-  );
-};
-export default ViewEditScheduledTasks;
